Copy-paste to make imminent refactoring easier
