+++ jib-maven-plugin/src/main/java/com/google/cloud/tools/jib/maven/BuildTarMojo.java
@@ -16,6 +16,7 @@
 
 package com.google.cloud.tools.jib.maven;
 
+import com.google.cloud.tools.jib.JibLogger;
 import com.google.cloud.tools.jib.cache.CacheDirectoryCreationException;
 import com.google.cloud.tools.jib.configuration.BuildConfiguration;
 import com.google.cloud.tools.jib.configuration.CacheConfiguration;
@@ -28,13 +29,16 @@ import com.google.cloud.tools.jib.frontend.HelpfulSuggestions;
 import com.google.cloud.tools.jib.frontend.JavaEntrypointConstructor;
 import com.google.cloud.tools.jib.frontend.SystemPropertyValidator;
 import com.google.cloud.tools.jib.http.Authorization;
+import com.google.cloud.tools.jib.http.Authorizations;
 import com.google.cloud.tools.jib.image.ImageReference;
 import com.google.cloud.tools.jib.registry.RegistryClient;
 import com.google.cloud.tools.jib.registry.credentials.RegistryCredentials;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
+import com.google.common.base.Strings;
 import java.nio.file.Paths;
 import java.time.Instant;
+import javax.annotation.Nullable;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.ResolutionScope;
@@ -82,7 +86,13 @@ public class BuildTarMojo extends JibPluginConfiguration {
     MavenSettingsServerCredentials mavenSettingsServerCredentials =
         new MavenSettingsServerCredentials(
             Preconditions.checkNotNull(session).getSettings(), settingsDecrypter, mavenJibLogger);
-    Authorization fromAuthorization = getBaseImageAuth(mavenJibLogger);
+    Authorization fromAuthorization =
+        getImageAuth(
+            mavenJibLogger,
+            "from",
+            "jib.from.auth.username",
+            "jib.from.auth.password",
+            getBaseImageAuth());
     RegistryCredentials knownBaseRegistryCredentials =
         fromAuthorization != null
             ? new RegistryCredentials(
@@ -147,5 +157,76 @@ public class BuildTarMojo extends JibPluginConfiguration {
     } catch (CacheDirectoryCreationException | BuildStepsExecutionException ex) {
       throw new MojoExecutionException(ex.getMessage(), ex.getCause());
     }
+  }
+
+  /**
+   * Gets an {@link Authorization} from a username and password. First tries system properties, then
+   * tries build configuration, otherwise returns null.
+   *
+   * <p>TODO: Consolidate with the other mojos.
+   *
+   * @param logger the {@link JibLogger} used to print warnings messages
+   * @param imageProperty the image configuration's name (i.e. "from" or "to")
+   * @param usernameProperty the name of the username system property
+   * @param passwordProperty the name of the password system property
+   * @param auth the configured credentials
+   * @return a new {@link Authorization} from the system properties or build configuration, or
+   *     {@code null} if neither is configured.
+   */
+  @Nullable
+  private static Authorization getImageAuth(
+      JibLogger logger,
+      String imageProperty,
+      String usernameProperty,
+      String passwordProperty,
+      AuthConfiguration auth) {
+    // System property takes priority over build configuration
+    String commandlineUsername = System.getProperty(usernameProperty);
+    String commandlinePassword = System.getProperty(passwordProperty);
+    if (!Strings.isNullOrEmpty(commandlineUsername)
+        && !Strings.isNullOrEmpty(commandlinePassword)) {
+      return Authorizations.withBasicCredentials(commandlineUsername, commandlinePassword);
+    }
+
+    // Warn if a system property is missing
+    if (!Strings.isNullOrEmpty(commandlinePassword) && Strings.isNullOrEmpty(commandlineUsername)) {
+      logger.warn(
+          passwordProperty
+              + " system property is set, but "
+              + usernameProperty
+              + " is not; attempting other authentication methods.");
+    }
+    if (!Strings.isNullOrEmpty(commandlineUsername) && Strings.isNullOrEmpty(commandlinePassword)) {
+      logger.warn(
+          usernameProperty
+              + " system property is set, but "
+              + passwordProperty
+              + " is not; attempting other authentication methods.");
+    }
+
+    // Check auth configuration next; warn if they aren't both set
+    if (Strings.isNullOrEmpty(auth.getUsername()) && Strings.isNullOrEmpty(auth.getPassword())) {
+      return null;
+    }
+    if (Strings.isNullOrEmpty(auth.getUsername())) {
+      logger.warn(
+          "<"
+              + imageProperty
+              + "><auth><username> is missing from maven configuration; ignoring <"
+              + imageProperty
+              + "><auth> section.");
+      return null;
+    }
+    if (Strings.isNullOrEmpty(auth.getPassword())) {
+      logger.warn(
+          "<"
+              + imageProperty
+              + "><auth><password> is missing from maven configuration; ignoring <"
+              + imageProperty
+              + "><auth> section.");
+      return null;
+    }
+
+    return Authorizations.withBasicCredentials(auth.getUsername(), auth.getPassword());
   }
 }
